requirements.txt
boto==2.49.0
boto3==1.8.1
botocore==1.11.1
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
gensim==3.5.0
h5py==2.8.0
idna==2.7
jmespath==0.9.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
nltk==3.3
numpy==1.15.1
pandas==0.23.4
pyspellchecker==0.1.3
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.6.0
urllib3==1.23
